Refactored file system cmds
