+++ client/command/filesystem/README.md
@@ -2,4 +2,3 @@ Filesystem
 ==========
 
 This package implements file system commands such as `ls`, `cd`, `rm`, `download`, `upload`, etc.
-
